# Improve the performance of the Memory store