--- a/drivers/store/tests/tests.go
+++ b/drivers/store/tests/tests.go
@@ -145,28 +145,24 @@ func TestStoreConcurrentAccess(t *testing.T, store limiter.Store) {
 
 // BenchmarkStoreSequentialAccess executes a benchmark against a store without parallel setting.
 func BenchmarkStoreSequentialAccess(b *testing.B, store limiter.Store) {
-	is := require.New(b)
 	ctx := context.Background()
 
-	limiter := limiter.New(store, limiter.Rate{
+	instance := limiter.New(store, limiter.Rate{
 		Limit:  100000,
 		Period: 10 * time.Second,
 	})
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		lctx, err := limiter.Get(ctx, "foo")
-		is.NoError(err)
-		is.NotZero(lctx)
+		_, _ = instance.Get(ctx, "foo")
 	}
 }
 
 // BenchmarkStoreConcurrentAccess executes a benchmark against a store with parallel setting.
 func BenchmarkStoreConcurrentAccess(b *testing.B, store limiter.Store) {
-	is := require.New(b)
 	ctx := context.Background()
 
-	limiter := limiter.New(store, limiter.Rate{
+	instance := limiter.New(store, limiter.Rate{
 		Limit:  100000,
 		Period: 10 * time.Second,
 	})
@@ -174,9 +170,7 @@ func BenchmarkStoreConcurrentAccess(b *testing.B, store limiter.Store) {
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			lctx, err := limiter.Get(ctx, "foo")
-			is.NoError(err)
-			is.NotZero(lctx)
+			_, _ = instance.Get(ctx, "foo")
 		}
 	})
 }