This pull request tries to tackle performance issues reported with the memory store (It should close #86).

For comparison, this is the benchmark on master branch:

```
➤ go test -v -benchtime=40s -run=XXX -benchmem -bench .
goos: linux
goarch: amd64
pkg: github.com/ulule/limiter/v3/drivers/store/memory
BenchmarkRedisStoreSequentialAccess
BenchmarkRedisStoreSequentialAccess-4   	106805974	      396 ns/op	     80 B/op	      3 allocs/op
BenchmarkRedisStoreConcurrentAccess
BenchmarkRedisStoreConcurrentAccess-4   	127844050	      398 ns/op	     80 B/op	      3 allocs/op
PASS
```

And the benchmark with these improvements:

```
➤ go test -v -benchtime=40s -run=XXX -benchmem -bench .
goos: linux
goarch: amd64
pkg: github.com/ulule/limiter/v3/drivers/store/memory
BenchmarkMemoryStoreSequentialAccess
BenchmarkMemoryStoreSequentialAccess-4   	167021902	      289 ns/op	      0 B/op	      0 allocs/op
BenchmarkMemoryStoreConcurrentAccess
BenchmarkMemoryStoreConcurrentAccess-4   	182060744	      309 ns/op	      0 B/op	      0 allocs/op
PASS
```